requirements.txt
streamlit
supabase
pandas
numpy
scikit-learn
matplotlib
python-dotenv
